Making a few refactors to simplify the components
