+++ frontend/src/components/PokemonCardModal/CardAttacks.js
@@ -1,12 +1,13 @@
 import React from "react";
-import styled, { withTheme } from "styled-components";
+import styled from "styled-components";
+import TypeBadge from "./TypeBadge";
 
 const Section = styled.div`
   margin-bottom: ${(props) => props.theme.spacing.xl};
 `;
 
 const SectionTitle = styled.h3`
-  margin: 0 0 10px 0;
+  margin-bottom: 10px;
   font-size: ${(props) => props.theme.fontSizes.lg};
   color: ${(props) => props.theme.colors.primary};
 `;
@@ -43,16 +44,6 @@ const CostContainer = styled.div`
   margin-bottom: ${(props) => props.theme.spacing.sm};
 `;
 
-const TypeBadge = styled.span`
-  background-color: ${(props) => props.bgColor};
-  color: ${(props) => props.theme.colors.white};
-  padding: ${(props) => props.theme.spacing.xs}
-    ${(props) => props.theme.spacing.sm};
-  border-radius: ${(props) => props.theme.spacing.md};
-  font-size: ${(props) => props.theme.fontSizes.xs};
-  font-weight: bold;
-`;
-
 const AttackEffect = styled.p`
   margin: 0;
   font-size: ${(props) => props.theme.fontSizes.sm};
@@ -60,12 +51,7 @@ const AttackEffect = styled.p`
   font-style: italic;
 `;
 
-// Helper function to get type color from theme
-const getTypeColor = (type, theme) => {
-  return theme.typeColors[type] || theme.typeColors.default;
-};
-
-const CardAttacks = ({ attacks, theme }) => {
+const CardAttacks = ({ attacks }) => {
   if (!attacks || attacks.length === 0) {
     return null;
   }
@@ -82,9 +68,7 @@ const CardAttacks = ({ attacks, theme }) => {
           {attack.cost && attack.cost.length > 0 && (
             <CostContainer>
               {attack.cost.map((cost, costIndex) => (
-                <TypeBadge key={costIndex} bgColor={getTypeColor(cost, theme)}>
-                  {cost}
-                </TypeBadge>
+                <TypeBadge key={costIndex} type={cost} />
               ))}
             </CostContainer>
           )}
@@ -95,4 +79,4 @@ const CardAttacks = ({ attacks, theme }) => {
   );
 };
 
-export default withTheme(CardAttacks);
+export default CardAttacks;
