+++ frontend/src/components/PokemonCardsList/SearchBar.js
@@ -1,5 +1,6 @@
 import React, { useState } from "react";
 import styled from "styled-components";
+import Button from "./Button";
 
 const SearchForm = styled.form`
   margin-bottom: ${(props) => props.theme.spacing.xl};
@@ -15,28 +16,6 @@ const SearchInput = styled.input`
 
   &:disabled {
     background-color: ${(props) => props.theme.colors.lightGray};
-    cursor: not-allowed;
-  }
-`;
-
-const Button = styled.button`
-  padding: ${(props) => props.theme.spacing.sm}
-    ${(props) => props.theme.spacing.md};
-  margin-left: ${(props) => props.marginLeft || "0"};
-  background-color: ${(props) => props.theme.colors.primary};
-  color: ${(props) => props.theme.colors.white};
-  border: none;
-  border-radius: ${(props) => props.theme.spacing.xs};
-  cursor: pointer;
-  font-size: ${(props) => props.theme.fontSizes.base};
-
-  &:hover {
-    background-color: #555;
-  }
-
-  &:disabled {
-    background-color: ${(props) => props.theme.colors.lightGray};
-    color: ${(props) => props.theme.colors.secondary};
     cursor: not-allowed;
   }
 `;
@@ -77,7 +56,7 @@ const SearchBar = ({ onSearch, disabled }) => {
             setSearchTerm("");
             onSearch("");
           }}
-          marginLeft="10px"
+          style={{ marginLeft: "10px" }}
         >
           Clear
         </Button>
